+++ resource/addr.go
@@ -10,6 +10,7 @@ import (
 )
 
 type Addr struct {
+	DiscoveryMeta `yaml:",inline" json:",inline"`
 	Title        string  `json:"title,omitempty" yaml:"title,omitempty"`
 	Meta         meta    `json:"meta,omitempty" yaml:"meta,omitempty"`
 	id           string  `json:"-" yaml:"-"`