+++ resource/command.go
@@ -13,6 +13,7 @@ import (
 )
 
 type Command struct {
+	DiscoveryMeta `yaml:",inline" json:",inline"`
 	Title      string  `json:"title,omitempty" yaml:"title,omitempty"`
 	Meta       meta    `json:"meta,omitempty" yaml:"meta,omitempty"`
 	id         string  `json:"-" yaml:"-"`