+++ resource/dns.go
@@ -11,6 +11,7 @@ import (
 )
 
 type DNS struct {
+	DiscoveryMeta `yaml:",inline" json:",inline"`
 	Title       string  `json:"title,omitempty" yaml:"title,omitempty"`
 	Meta        meta    `json:"meta,omitempty" yaml:"meta,omitempty"`
 	id          string  `json:"-" yaml:"-"`