+++ resource/gossfile.go
@@ -6,6 +6,7 @@ import (
 )
 
 type Gossfile struct {
+	DiscoveryMeta `yaml:",inline" json:",inline"`
 	Title string `json:"title,omitempty" yaml:"title,omitempty"`
 	Meta  meta   `json:"meta,omitempty" yaml:"meta,omitempty"`
 	Path  string `json:"-" yaml:"-"`